add benchmark validate request
